Full project restructure completed
