+++ src/core/components/Certificate.tsx
