+++ src/core/components/PhotoViewer.tsx
