+++ src/core/partials/PostContent.tsx
