+++ src/layouts/Base.astro
@@ -30,7 +30,9 @@ const {
   <body class="bg-slate-900 text-gray-100 px-5">
     <div class="container mx-auto pt-10 pb-10 max-w-4xl px-5">
       <slot />
-      <Footer />
+      <footer class="mt-12">
+        <Footer />
+      </footer>
     </div>
   </body>
 </html>@/components/core/partials/Footer
