+++ src/pages/buy.co.sz-website.mdx
@@ -25,7 +25,7 @@ screnshots:
 preview_ids: ["images"]
 ---
 
-import PhotoViewer from "@/core/partials/PhotoViewer";
+import PhotoViewer from "@/core/components/PhotoViewer";
 
 ## Technical Overview
